% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{expand_som_spec}
\alias{expand_som_spec}
\title{Expand an ensemble specification into its model budget}
\usage{
expand_som_spec(spec)
}
\arguments{
\item{spec}{A \code{som_spec} object.}
}
\value{
A data frame with one row per grid and random seed.
}
\description{
Expand an ensemble specification into its model budget
}
