% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.som_audit}
\alias{plot.som_audit}
\title{Plot SOM audit trade-offs}
\usage{
\method{plot}{som_audit}(x, ...)
}
\arguments{
\item{x}{A \code{som_audit} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot SOM audit trade-offs
}
