% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.som_cross_comparison}
\alias{plot.som_cross_comparison}
\title{Plot cross-model agreement without ranking methods}
\usage{
\method{plot}{som_cross_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{som_cross_comparison} object.}

\item{...}{Unused.}
}
\value{
A faceted \code{ggplot} object showing ARI and AMI distributions.
}
\description{
Plot cross-model agreement without ranking methods
}
