% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.som_newdata_mapping}
\alias{plot.som_newdata_mapping}
\title{Plot new-data mapping shift across an ensemble}
\usage{
\method{plot}{som_newdata_mapping}(x, ...)
}
\arguments{
\item{x}{A \code{som_newdata_mapping} object.}

\item{...}{Unused.}
}
\value{
An editable \code{ggplot} object.
}
\description{
Plot new-data mapping shift across an ensemble
}
