% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.som_transfer_audit}
\alias{plot.som_transfer_audit}
\title{Plot mapping shift in held-out domains}
\usage{
\method{plot}{som_transfer_audit}(x, ...)
}
\arguments{
\item{x}{A \code{som_transfer_audit} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot mapping shift in held-out domains
}
